requirements.txt
matplotlib
python-dotenv
pytest
pandas

streamlit
streamlit_folium

pyspark